+++ R/mltools.R
@@ -52,7 +52,7 @@ autostrat <- function(x,
   if (!is.ichimoku(x)) stop("autostrat() only works on ichimoku objects", call. = FALSE)
   dir <- match.arg(dir)
   if (!level %in% 1:3) {
-    warning("Specified 'level' invalid - falling back to default of 1", call. = FALSE)
+    warning("Specified 'level' invalid - reverting to default of 1", call. = FALSE)
     level <- 1
   }